+++ plugins/scaffolder-backend-module-github/src/actions/githubRepoCreate.ts
@@ -41,118 +41,50 @@ export function createGithubRepoCreateAction(options: {
 }) {
   const { integrations, githubCredentialsProvider } = options;
 
-  return createTemplateAction<{
-    repoUrl: string;
-    description?: string;
-    homepage?: string;
-    access?: string;
-    deleteBranchOnMerge?: boolean;
-    gitAuthorName?: string;
-    gitAuthorEmail?: string;
-    allowRebaseMerge?: boolean;
-    allowSquashMerge?: boolean;
-    squashMergeCommitTitle?: 'PR_TITLE' | 'COMMIT_OR_PR_TITLE';
-    squashMergeCommitMessage?: 'PR_BODY' | 'COMMIT_MESSAGES' | 'BLANK';
-    allowMergeCommit?: boolean;
-    allowAutoMerge?: boolean;
-    allowUpdateBranch?: boolean;
-    requireCodeOwnerReviews?: boolean;
-    bypassPullRequestAllowances?: {
-      users?: string[];
-      teams?: string[];
-      apps?: string[];
-    };
-    requiredApprovingReviewCount?: number;
-    restrictions?: {
-      users: string[];
-      teams: string[];
-      apps?: string[];
-    };
-    requiredStatusCheckContexts?: string[];
-    requireBranchesToBeUpToDate?: boolean;
-    requiredConversationResolution?: boolean;
-    repoVisibility?: 'private' | 'internal' | 'public';
-    collaborators?: Array<
-      | {
-          user: string;
-          access: string;
-        }
-      | {
-          team: string;
-          access: string;
-        }
-      | {
-          /** @deprecated This field is deprecated in favor of team */
-          username: string;
-          access: 'pull' | 'push' | 'admin' | 'maintain' | 'triage';
-        }
-    >;
-    hasProjects?: boolean;
-    hasWiki?: boolean;
-    hasIssues?: boolean;
-    token?: string;
-    topics?: string[];
-    repoVariables?: { [key: string]: string };
-    secrets?: { [key: string]: string };
-    oidcCustomization?: {
-      useDefault: boolean;
-      includeClaimKeys?: string[];
-    };
-    requireCommitSigning?: boolean;
-    requiredLinearHistory?: boolean;
-    customProperties?: { [key: string]: string };
-    subscribe?: boolean;
-  }>({
+  return createTemplateAction({
     id: 'github:repo:create',
     description: 'Creates a GitHub repository.',
     examples,
     schema: {
       input: {
-        type: 'object',
-        required: ['repoUrl'],
-        properties: {
-          repoUrl: inputProps.repoUrl,
-          description: inputProps.description,
-          homepage: inputProps.homepage,
-          access: inputProps.access,
-          requireCodeOwnerReviews: inputProps.requireCodeOwnerReviews,
-          bypassPullRequestAllowances: inputProps.bypassPullRequestAllowances,
-          requiredApprovingReviewCount: inputProps.requiredApprovingReviewCount,
-          restrictions: inputProps.restrictions,
-          requiredStatusCheckContexts: inputProps.requiredStatusCheckContexts,
-          requireBranchesToBeUpToDate: inputProps.requireBranchesToBeUpToDate,
-          requiredConversationResolution:
-            inputProps.requiredConversationResolution,
-          repoVisibility: inputProps.repoVisibility,
-          deleteBranchOnMerge: inputProps.deleteBranchOnMerge,
-          allowMergeCommit: inputProps.allowMergeCommit,
-          allowSquashMerge: inputProps.allowSquashMerge,
-          squashMergeCommitTitle: inputProps.squashMergeCommitTitle,
-          squashMergeCommitMessage: inputProps.squashMergeCommitMessage,
-          allowRebaseMerge: inputProps.allowRebaseMerge,
-          allowAutoMerge: inputProps.allowAutoMerge,
-          allowUpdateBranch: inputProps.allowUpdateBranch,
-          collaborators: inputProps.collaborators,
-          hasProjects: inputProps.hasProjects,
-          hasWiki: inputProps.hasWiki,
-          hasIssues: inputProps.hasIssues,
-          token: inputProps.token,
-          topics: inputProps.topics,
-          repoVariables: inputProps.repoVariables,
-          secrets: inputProps.secrets,
-          oidcCustomization: inputProps.oidcCustomization,
-          requiredCommitSigning: inputProps.requiredCommitSigning,
-          requiredLinearHistory: inputProps.requiredLinearHistory,
-          customProperties: inputProps.customProperties,
-          subscribe: inputProps.subscribe,
-        },
+        repoUrl: inputProps.repoUrl,
+        description: inputProps.description,
+        homepage: inputProps.homepage,
+        access: inputProps.access,
+        requireCodeOwnerReviews: inputProps.requireCodeOwnerReviews,
+        bypassPullRequestAllowances: inputProps.bypassPullRequestAllowances,
+        requiredApprovingReviewCount: inputProps.requiredApprovingReviewCount,
+        restrictions: inputProps.restrictions,
+        requiredStatusCheckContexts: inputProps.requiredStatusCheckContexts,
+        requireBranchesToBeUpToDate: inputProps.requireBranchesToBeUpToDate,
+        requiredConversationResolution:
+          inputProps.requiredConversationResolution,
+        repoVisibility: inputProps.repoVisibility,
+        deleteBranchOnMerge: inputProps.deleteBranchOnMerge,
+        allowMergeCommit: inputProps.allowMergeCommit,
+        allowSquashMerge: inputProps.allowSquashMerge,
+        squashMergeCommitTitle: inputProps.squashMergeCommitTitle,
+        squashMergeCommitMessage: inputProps.squashMergeCommitMessage,
+        allowRebaseMerge: inputProps.allowRebaseMerge,
+        allowAutoMerge: inputProps.allowAutoMerge,
+        allowUpdateBranch: inputProps.allowUpdateBranch,
+        collaborators: inputProps.collaborators,
+        hasProjects: inputProps.hasProjects,
+        hasWiki: inputProps.hasWiki,
+        hasIssues: inputProps.hasIssues,
+        token: inputProps.token,
+        topics: inputProps.topics,
+        repoVariables: inputProps.repoVariables,
+        secrets: inputProps.secrets,
+        oidcCustomization: inputProps.oidcCustomization,
+        requiredCommitSigning: inputProps.requiredCommitSigning,
+        requiredLinearHistory: inputProps.requiredLinearHistory,
+        customProperties: inputProps.customProperties,
+        subscribe: inputProps.subscribe,
       },
       output: {
-        type: 'object',
-        properties: {
-          remoteUrl: outputProps.remoteUrl,
-          repoContentsUrl: outputProps.repoContentsUrl,
-        },
+        remoteUrl: outputProps.remoteUrl,
+        repoContentsUrl: outputProps.repoContentsUrl,
       },
     },
     async handler(ctx) {