+++ plugins/scaffolder-backend-module-gitlab/src/actions/gitlabProjectDeployTokenCreate.ts
@@ -18,8 +18,6 @@ import { InputError } from '@backstage/errors';
 import { ScmIntegrationRegistry } from '@backstage/integration';
 import { createTemplateAction } from '@backstage/plugin-scaffolder-node';
 import { DeployTokenScope, Gitlab } from '@gitbeaker/rest';
-import { z } from 'zod';
-import commonGitlabConfig from '../commonGitlabConfig';
 import { getToken } from '../util';
 import { examples } from './gitlabProjectDeployTokenCreate.examples';
 
@@ -37,22 +35,46 @@ export const createGitlabProjectDeployTokenAction = (options: {
     id: 'gitlab:projectDeployToken:create',
     examples,
     schema: {
-      input: commonGitlabConfig.merge(
-        z.object({
-          projectId: z.union([z.number(), z.string()], {
+      input: {
+        repoUrl: z =>
+          z.string({
+            description: `Accepts the format 'gitlab.com?repo=project_name&owner=group_name' where 'project_name' is the repository name and 'group_name' is a group or username`,
+          }),
+        token: z =>
+          z
+            .string({
+              description: 'The token to use for authorization to GitLab',
+            })
+            .optional(),
+        projectId: z =>
+          z.union([z.number(), z.string()], {
             description: 'Project ID',
           }),
-          name: z.string({ description: 'Deploy Token Name' }),
-          username: z
-            .string({ description: 'Deploy Token Username' })
+        name: z =>
+          z.string({
+            description: 'Deploy Token Name',
+          }),
+        username: z =>
+          z
+            .string({
+              description: 'Deploy Token Username',
+            })
             .optional(),
-          scopes: z.array(z.string(), { description: 'Scopes' }),
-        }),
-      ),
-      output: z.object({
-        deploy_token: z.string({ description: 'Deploy Token' }),
-        user: z.string({ description: 'User' }),
-      }),
+        scopes: z =>
+          z.array(z.string(), {
+            description: 'Scopes',
+          }),
+      },
+      output: {
+        deploy_token: z =>
+          z.string({
+            description: 'Deploy Token',
+          }),
+        user: z =>
+          z.string({
+            description: 'User',
+          }),
+      },
     },
     async handler(ctx) {
       ctx.logger.info(`Creating Token for Project "${ctx.input.projectId}"`);