+++ plugins/notifications/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-notifications
 
+## 0.5.6-next.1
+
+### Patch Changes
+
+- 1fb5f06: Adds ability for user to turn on/off notifications for specific topics within an origin.
+- Updated dependencies
+  - @backstage/plugin-notifications-common@0.0.9-next.0
+  - @backstage/core-compat-api@0.4.3-next.1
+  - @backstage/core-components@0.17.2
+  - @backstage/core-plugin-api@1.10.7
+  - @backstage/errors@1.2.7
+  - @backstage/frontend-plugin-api@0.10.3-next.0
+  - @backstage/theme@0.6.6
+  - @backstage/types@1.2.1
+  - @backstage/plugin-signals-react@0.0.13
+
 ## 0.5.6-next.0
 
 ### Patch Changes