+++ plugins/user-settings/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @backstage/plugin-user-settings
 
+## 0.8.23-next.1
+
+### Patch Changes
+
+- ab6d3dd: Add some translation to user-setting plugin
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.18.1-next.1
+  - @backstage/catalog-model@1.7.4
+  - @backstage/core-app-api@1.17.0
+  - @backstage/core-compat-api@0.4.3-next.1
+  - @backstage/core-components@0.17.2
+  - @backstage/core-plugin-api@1.10.7
+  - @backstage/errors@1.2.7
+  - @backstage/frontend-plugin-api@0.10.3-next.0
+  - @backstage/theme@0.6.6
+  - @backstage/types@1.2.1
+  - @backstage/plugin-signals-react@0.0.13
+  - @backstage/plugin-user-settings-common@0.0.1
+
 ## 0.8.23-next.0
 
 ### Patch Changes